server/requirements.txt
flask==3.0.0
cryptography==41.0.7
python-dotenv==1.0.0
influxdb-client==1.38.0
